% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrpca.R
\name{qrpca}
\alias{qrpca}
\title{QR based PCA}
\usage{
qrpca(x, center = TRUE, scale = FALSE, cuda = FALSE)
}
\arguments{
\item{x}{a numeric matrix or data frame which provides the data
for the principal component analysis.}

\item{center}{a logical value indicating whether the empirical mean of (the
columns) of \code{x} should be subtracted.}

\item{scale}{a logical value indicating whether the columns of \code{x}
 should be scaled to have unit variance before the analysis takes place. The
 default is \code{FALSE} for consistency with \code{prcomp}.
@param cuda a logical value indicating whether cuda acceleration should be
used. The default is \code{FALSE}.}
}
\value{
\code{qrpca} returns a list with class \code{prcomp}
  containing the following elements: \item{sdev}{the additional standard
  deviation explained by each component.}
  \item{rotation}{the matrix of loadings,
  containing the principal axes as columns.}
  \item{x}{the scores matrix
   containing the principal components as columns.}
  \item{center, scale}{the centering and
  scaling used, or \code{FALSE}}
}
\description{
Performs a QR based principal component analysis,
The result is returned as an object of class \code{prcomp}.
}
\details{
\code{qrpca} computes a principal component (PC) using QR decomposition as
intermediate step. Matrix operations mostly uses  \code{torch} for
performance.
}
